+++ src/parse.jl
@@ -15,7 +15,7 @@ function parse_commands(d::Dict)
     for typ_key in (ANALOG_KEY, DIGITAL_KEY)
         for k in keys(d[typ_key])
             v = d[typ_key][k]
-            push!(output, ImagineCommand(rig, k, convert(RLEVector, v), d[COMPONENT_KEY], Int64(d[METADATA_KEY]["sample rate"])))
+            push!(output, ImagineCommand(rig, k, convert(RLEVector, v), d[COMPONENT_KEY], Int(d[METADATA_KEY]["sample rate"])))
         end
     end
     return output
@@ -38,7 +38,7 @@ function parse_command(d::Dict, comname::String)
     for typ_key in (ANALOG_KEY, DIGITAL_KEY)
         ad = d[typ_key]
         if haskey(ad, comname)
-            return ImagineCommand(rig, comname, convert(RLEVector, ad[comname]), d[COMPONENT_KEY], Int64(d[METADATA_KEY]["sample rate"]))
+            return ImagineCommand(rig, comname, convert(RLEVector, ad[comname]), d[COMPONENT_KEY], Int(d[METADATA_KEY]["sample rate"]))
         end
     end
     error("Command signal name not found")
